change type into float

+++ data_loader_home.py
@@ -42,7 +42,7 @@ def load_dataset(N=1,NP=4000,folder='../data/simple/',s=0):
     obs = (obs - lower) / bound - 1.0
     print('after normalization:')
     print(obs)
-    obs = obs.reshape(N,-1)
+    obs = obs.reshape(N,-1).astype(np.float32)
 
 
     ## calculating length of the longest trajectory
@@ -129,7 +129,7 @@ def load_test_dataset(N=100,NP=200, s=0,sp=4000, folder='../data/simple/'):
         obs = (obs - lower) / bound - 1.0
         print('after normalization:')
         print(obs)
-        obs = obs.reshape(N, -1)
+        obs = obs.reshape(N, -1).astype(np.float32)
     ## calculating length of the longest trajectory
     max_length=0
     path_lengths=np.zeros((N,NP),dtype=np.int8)
